pytest.ini: Switch to importlib for imports
When running the testsuite with vagrant installed, the
test_command_init_scenario() is using the test/vagrant
class instead of the python-vagrant module. This leads
to an error about the Vagrant() not existing.
Switching to importlib seems to solve the issue but
was breaking gce test suite since it was relying on
the test file name to find the driver name.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>