tox.ini: Install ansible for tests
molecule is using ansible-compat, which calls ansible-config.
With current setup, when run inside a sandbox with nothing
installed, ansible-compat use will lead to an error since
nothing is installing ansible (ansible-config).
It's tempting to say it's a bug in ansible-compat but in
the mean time, just install ansible.

Note:
This is not detected by the CI since the github workflow is
installing ansible.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>